Fix gas query...
The "standard" QueryBuilder query was not built with sparse data in mind.

In this case, gas usage is only registered when it changes,
while, for example, power is constantly being used in the house, so it is
(almost) guaranteed that there are measurements for every moment of the day.

In the case of gas usage, the first measurement of the day was not taken into
account since that was used as the "start" measurement.

In the new situation, the last measurement of the previous period (day, month, year)
is included in the query. This fixes this problem.