Migrate from influxdb to postgres + timescale...
I ran into a number of problems with influx:

1) Suddenly getting null pointer exceptions in Influx

These exceptions occur in the influx engine for data in some time spans

This could be a problem with the installation, since I _am_ able to do
those queries by restoring a backup to a different installation in a VM.

I don't want to spend time debugging this and so have no confidence that
it could also start happening in the VM, which would mean that I lose my
data.

2) The Flux query language being deprecated

I don't want to rewrite all my queries when upgrading to version 3 at
some point in the future with the NPEs mentioned above.

Additionally, when exporting the data

3) Problems extracting data through Flux

Even though I could read all the data on the backup VM, I could not use
the Flux language to extract the data:

It failed on the GROUP BY(duration) clause, even though it should have
been perfectly valid.

So to extract the data, I needed to use InfluxQL. Not a big problem, but
suspicious again.

4) Not all data is visible when querying through InfluxQL instead of Flux.

I could not get the temperature data out through InfluxQL since those
measurements were simply not visible. Some other measurements in the same
bucket _were_ visible, but not those.

So for this, I needed to use Flux again, through which the data _was_
visible.

Last point is the fact that it's a non-standard database implementation,
so if anything goes wrong, I don't expect there to be any tooling to fix
the data.

Contrast this with Postgres which has been in use for decades.

I don't expect it to be easy to read the Influx backup files to restore
data, if push comes to shove.

In conclusion: I don't have enough confidence to keep using Influx.

It does seem like my initial implementation with Timescale is a bit
slower than Influx, but I haven't looked at things like continuous
aggregates yet. That might make everything snappier. For the rest,
caching is my friend :)